+++ src/datastructures.jl
@@ -1,4 +1,33 @@
 """
+    AbstractGUIobj
+
+Supertype for EnergyModelsGUI objects representing `Node`s/`Link`s/`Area`s/`Transmission`s.
+"""
+abstract type AbstractGUIobj end
+
+"""
+    ProcInvData
+
+Type for storing processed investment data.
+
+# Fields
+
+- **`inv_times::Vector{String}`** a vector of formatted strings for added investments.
+- **`capex::Vector{Number}`** contains the capex of all times with added investments.
+- **`invested::Bool`** indicates if the element has been invested in.
+"""
+struct ProcInvData
+    inv_times::Vector{String}
+    capex::Vector{Number}
+    invested::Bool
+end
+function ProcInvData()
+    return ProcInvData(String[], Number[], false)
+end
+
+"""
+    EnergySystemDesign <: AbstractGUIobj
+
 Mutable type for providing a flexible data structure for modeling and working with complex
 energy system designs in Julia.
 
@@ -22,8 +51,9 @@ energy system designs in Julia.
 - **`file::String`** is the filename or path associated with the `EnergySystemDesign`.
 - **`plots::Vector{Any}`** is a vector with all Makie object associated with this object.
   The value does not have to be provided.
+- **`invest_data::ProcInvData`** stores processed investment data.
 """
-mutable struct EnergySystemDesign
+mutable struct EnergySystemDesign <: AbstractGUIobj
     parent::Union{Symbol,Nothing}
     system::Dict
     id_to_color_map::Dict
@@ -35,6 +65,7 @@ mutable struct EnergySystemDesign
     color::Observable{Symbol}
     wall::Observable{Symbol}
     file::String
+    inv_data::ProcInvData
     plots::Vector{Any}
 end
 function EnergySystemDesign(
@@ -62,11 +93,14 @@ function EnergySystemDesign(
         color,
         wall,
         file,
+        ProcInvData(),
         Any[],
     )
 end
 
 """
+    Connection <: AbstractGUIobj
+
 Mutable type for providing a flexible data structure for connections between
 `EnergySystemDesign`s.
 
@@ -79,12 +113,14 @@ Mutable type for providing a flexible data structure for connections between
 - **`connection::Union{Link,Transmission}`** is the EMX connection structure.
 - **`colors::Vector{RGB}`** is the associated colors of the connection.
 - **`plots::Vector{Any}`** is a vector with all Makie object associated with this object.
+- **`invest_data::ProcInvData`** stores processed investment data.
 """
-mutable struct Connection
+mutable struct Connection <: AbstractGUIobj
     from::EnergySystemDesign
     to::EnergySystemDesign
     connection::Union{Link,Transmission}
     colors::Vector{RGB}
+    inv_data::ProcInvData
     plots::Vector{Any}
 end
 function Connection(
@@ -94,10 +130,29 @@ function Connection(
     id_to_color_map::Dict{Any,Any},
 )
     colors::Vector{RGB} = get_resource_colors(connection, id_to_color_map)
-    return Connection(from, to, connection, colors, Any[])
+    return Connection(from, to, connection, colors, ProcInvData(), Any[])
 end
 
 """
+    EnergySystemIterator
+
+Type for iterating over nested `EnergySystemDesign` structures, enabling
+recursion through `AbstractGUIobj`s.
+
+# Fields
+
+- **`stack::Vector{<:AbstractGUIobj}`** is the stack used to manage the iteration
+  through the nested `EnergySystemDesign` components (and its connections).
+  It starts with the initial `EnergySystemDesign` object and progressively includes its
+  subcomponents as the iteration proceeds.
+"""
+struct EnergySystemIterator
+    stack::Vector{AbstractGUIobj}
+end
+
+"""
+    GUI
+
 The main type for the realization of the GUI.
 
 # Fields
@@ -165,16 +220,26 @@ function Base.show(io::IO, obj::EnergySystemDesign)
     println(io, "  wall (Observable{Symbol}): ", obj.wall)
 
     println(io, "  file (String): ", obj.file)
-    return println(io, "  plots (Vector{Any}): ", obj.plots)
+    println(io, "  inv_data (ProcInvData): ", obj.inv_data)
+    println(io, "  plots (Vector{Any}): ", obj.plots)
 end
 
 """
-    show(io::IO, con::Connection)
+    show(io::IO, obj::Connection)
 
 Print a simplified overview of the fields of a Connection `obj`.
 """
-function Base.show(io::IO, con::Connection)
-    return dump(io, con; maxdepth=2)
+function Base.show(io::IO, obj::Connection)
+    return dump(io, obj; maxdepth=2)
+end
+
+"""
+    show(io::IO, obj::ProcInvData)
+
+Print a simplified overview of the fields of a ProcInvData `obj`.
+"""
+function Base.show(io::IO, obj::ProcInvData)
+    return dump(io, obj)
 end
 
 """
@@ -185,6 +250,38 @@ Print a simplified overview of the fields of a GUI `gui`.
 function Base.show(io::IO, gui::GUI)
     return dump(io, gui; maxdepth=1)
 end
+
+"""
+    iterate(iter::EnergySystemIterator)
+
+Initialize the iteration over an `EnergySystemIterator`, returning the first `EnergySystemDesign` object
+in the stack and the iterator itself. If the stack is empty, return `nothing`.
+"""
+function Base.iterate(iter::EnergySystemIterator)
+    isempty(iter.stack) && return nothing
+    current = pop!(iter.stack)
+    if isa(current, EnergySystemDesign)
+        push!(iter.stack, current.components...)  # Add the components to the stack
+        push!(iter.stack, current.connections...)  # Add the connections to the stack
+    end
+    return current, iter
+end
+
+"""
+    iterate(design::EnergySystemDesign)
+
+Initialize the iteration over an `EnergySystemDesign`, returning the first `EnergySystemDesign` object
+and the iterator itself. The iteration traverses through all nested components and connections.
+"""
+Base.iterate(design::EnergySystemDesign) = iterate(EnergySystemIterator([design]))
+
+"""
+    iterate(design::EnergySystemDesign, state)
+
+Initialize the iteration over an `EnergySystemDesign`, returning the first `EnergySystemDesign` object
+and the iterator itself. The iteration traverses through all nested components and connections.
+"""
+Base.iterate(::EnergySystemDesign, state) = iterate(state)
 
 """
     get_parent(design::EnergySystemDesign)
@@ -201,11 +298,15 @@ Returns the `system` field of a `EnergySystemDesign` `design`.
 get_system(design::EnergySystemDesign) = design.system
 
 """
-    get_system_node(design::EnergySystemDesign)
+    get_element(design::EnergySystemDesign)
 
 Returns the system node (i.e. availability node for areas) of a `EnergySystemDesign` `design`.
 """
-get_system_node(design::EnergySystemDesign) = design.system[:node]
+function get_element(design::EnergySystemDesign)
+    if !isnothing(design.parent)
+        return design.system[:node]
+    end
+end
 
 """
     get_components(design::EnergySystemDesign)
@@ -257,6 +358,13 @@ Returns the `file` field of a `EnergySystemDesign` `design`.
 get_file(design::EnergySystemDesign) = design.file
 
 """
+    get_inv_data(design::EnergySystemDesign)
+
+Returns the `inv_data` field of a `EnergySystemDesign` `design`.
+"""
+get_inv_data(design::EnergySystemDesign) = design.inv_data
+
+"""
     get_plots(design::EnergySystemDesign)
 
 Returns the `plots` field of a `EnergySystemDesign` `design`.
@@ -278,11 +386,11 @@ Returns the `to` field of a `Connection` `conn`.
 get_to(conn::Connection) = conn.to
 
 """
-    get_connection(conn::Connection)
+    get_element(conn::Connection)
 
-Returns the `connection` field of a `Connection` `conn`.
+Returns the assosiated `Transmission`/`Link` of conn
 """
-get_connection(conn::Connection) = conn.connection
+get_element(conn::Connection) = conn.connection
 
 """
     get_colors(conn::Connection)
@@ -292,11 +400,42 @@ Returns the `colors` field of a `Connection` `conn`.
 get_colors(conn::Connection) = conn.colors
 
 """
+    get_inv_data(design::Connection)
+
+Returns the `inv_data` field of a `Connection` `design`.
+"""
+get_inv_data(design::Connection) = design.inv_data
+
+"""
     get_plots(conn::Connection)
 
 Returns the `plots` field of a `Connection` `conn`.
 """
 get_plots(conn::Connection) = conn.plots
+
+"""
+    get_inv_times(data)
+
+Returns the `inv_times` field of a `ProcInvData`/`AbstractGUIobj` object `data`.
+"""
+get_inv_times(data::ProcInvData) = data.inv_times
+get_inv_times(design::AbstractGUIobj) = get_inv_times(get_inv_data(design))
+
+"""
+    get_capex(data)
+
+Returns the `capex` of the investments of a `ProcInvData`/`AbstractGUIobj` object `data`.
+"""
+get_capex(data::ProcInvData) = data.capex
+get_capex(design::AbstractGUIobj) = get_capex(get_inv_data(design))
+
+"""
+    has_invested(data)
+
+Returns a boolean indicator if investment has occured.
+"""
+has_invested(data::ProcInvData) = data.invested
+has_invested(design::AbstractGUIobj) = has_invested(get_inv_data(design))
 
 """
     get_fig(gui::GUI)